+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Path Server will be documented in this file.
 ## Unreleased
 
 ### Added
+- Added language support assert to tree-sitter parser tests.
 - **Core**: Added MDX language support. ([#43](https://github.com/kunlinglio/path-server/pull/43))
 
 ## [1.3.1] - 2026-05-29